requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask==3.0.1
pypdf
langchain-community

-e .